LayoutTests/imported/w3c/web-platform-tests/css/css-break/table/table-parts-offsetheight-expected.txt


FAIL table assert_equals: expected 70 but got 120
FAIL colgroup assert_equals: expected 70 but got 114
FAIL col assert_equals: expected 70 but got 114
FAIL rowgroup assert_equals: expected 70 but got 114
FAIL row assert_equals: expected 70 but got 100
FAIL cell assert_equals: expected 70 but got 100
FAIL content assert_equals: expected 70 but got 100
